requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
httpie==2.3.0
idna==2.10
psycopg2==2.8.6
Pygments==2.7.2
PyJWT==1.7.1
PySocks==1.7.1
pytz==2020.4
requests==2.25.0
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.2
